+++ src/vector/product/fill.jl
@@ -5,6 +5,8 @@
 # For example when N = 10, the optimised path here brings from_linked_vec(d)(y) down 
 # from 731 ns to 59 ns, and to_linked_vec(d)(x) from 278 ns to 75 ns.
 
+using FillArrays: FillArrays
+
 """
     Elementwise{T,S}
 
@@ -38,23 +40,61 @@ end
 
 _map_inverse(t::Elementwise) = Elementwise(inverse(t.value), t.size)
 
-# Trait: returns true when the vector bijector for a distribution type is determined
-# solely by the type (not runtime parameter values).
-_has_constant_vec_bijector(::Type) = false
-_has_constant_vec_bijector(::Type{<:IDENTITY_UNIVARIATES}) = true
-_has_constant_vec_bijector(::Type{<:POSITIVE_UNIVARIATES}) = true
+"""
+    Bijectors.VectorBijectors.has_constant_vec_bijector(::Type{T}) where {T}
+
+Return `true` if the vector bijector for each element of a collection of distributions is
+determined solely by the type of the distribution, and not by any runtime parameter
+values.
+
+This is slightly confusing, so is best explained by example. Consider
+
+```julia
+d = product_distribution(array_of_dists)
+```
+
+If it can be inferred from `typeof(array_of_dists)` that each distribution inside
+`array_of_dists` has the same vector bijector, then
+`has_constant_vec_bijector(typeof(array_of_dists))` should return `true`.
+
+For example, if `array_of_dists` is a `FillArrays.Fill` of some distribution type, then we
+know that each distribution inside is the same, and so they all have the same vector
+bijector. Thus, we have that
+
+```julia
+has_constant_vec_bijector(::Type{<:FillArrays.Fill}) == true
+```
+
+For generic `AbstractArray`s or `Tuple`s, this will dispatch on the element type of the
+array. That means that if a `dist::D` (where `D<:Distribution`) has a constant vector
+bijector, we can simply mark `has_constant_vec_bijector(::Type{D}) == true`.
+
+For example, `Beta` has a constant vector bijector, because its support is always between 0
+and 1, regardless of its parameters.
+
+On the other hand, `Uniform` does not have a constant vector bijector, because its support
+depends on its parameters.
+"""
+has_constant_vec_bijector(::Type{TFill}) where {TFill<:FillArrays.Fill} = true
+function has_constant_vec_bijector(::Type{<:AbstractArray{T}}) where {T}
+    return has_constant_vec_bijector(T)
+end
+has_constant_vec_bijector(t::Type{<:Tuple}) = has_constant_vec_bijector(eltype(t))
+has_constant_vec_bijector(::Type) = false
+has_constant_vec_bijector(::Type{<:IDENTITY_UNIVARIATES}) = true
+has_constant_vec_bijector(::Type{<:POSITIVE_UNIVARIATES}) = true
 # between 0 and 1
-function _has_constant_vec_bijector(
+function has_constant_vec_bijector(
     ::Type{<:Union{D.Beta,D.KSOneSided,D.NoncentralBeta,D.LogitNormal}}
 )
     return true
 end
-_has_constant_vec_bijector(::Type{<:D.DiscreteUnivariateDistribution}) = true
+has_constant_vec_bijector(::Type{<:D.DiscreteUnivariateDistribution}) = true
 # Multivariates
-_has_constant_vec_bijector(::Type{<:D.AbstractMvNormal}) = true
-_has_constant_vec_bijector(::Type{<:D.AbstractMvLogNormal}) = true
-_has_constant_vec_bijector(::Type{<:SIMPLEX_MULTIVARIATES}) = true
-_has_constant_vec_bijector(::Type{<:D.DiscreteMultivariateDistribution}) = true
+has_constant_vec_bijector(::Type{<:D.AbstractMvNormal}) = true
+has_constant_vec_bijector(::Type{<:D.AbstractMvLogNormal}) = true
+has_constant_vec_bijector(::Type{<:SIMPLEX_MULTIVARIATES}) = true
+has_constant_vec_bijector(::Type{<:D.DiscreteMultivariateDistribution}) = true
 
 function (t::ProductVecTransform{<:Elementwise{F,Dims{M}},Nothing,Dims{N}})(
     x::AbstractArray{T}
@@ -66,6 +106,15 @@ function (t::ProductVecTransform{<:Elementwise{F,Dims{M}},Nothing,Dims{N}})(
         dims = ntuple(i -> i + N, Val(M))
         vec(stack(trf, eachslice(x; dims=dims)))
     end
+end
+function (::ProductVecTransform{<:Elementwise{TypedIdentity},Nothing,Dims{0}})(
+    x::AbstractArray
+)
+    # If the wrapped transform is TypedIdentity, and the distribution is univariate (i.e.,
+    # the 'base size' is `()::Dims{0}`), then the entire vectorisation transform amounts to
+    # just `vec`. This special case is hit for things like
+    # product_distribution(fill(Cauchy(), m1, m2, ...)).
+    return vec(x)
 end
 
 # Tiny struct that allows us to use map(eachslices) directly instead of a manual loop inside
@@ -102,6 +151,11 @@ function with_logabsdet_jacobian(
         y, lj.logjac
     end
 end
+function with_logabsdet_jacobian(
+    ::ProductVecTransform{<:Elementwise{TypedIdentity},Nothing,Dims{0}}, x::AbstractArray{T}
+) where {T}
+    return vec(x), _fzero(T)
+end
 
 function (t::ProductVecInvTransform{<:Elementwise{F,Dims{M}},Nothing,Dims{N}})(
     y::AbstractVector{T}
@@ -116,6 +170,11 @@ function (t::ProductVecInvTransform{<:Elementwise{F,Dims{M}},Nothing,Dims{N}})(
         dims = ntuple(i -> i + 1, Val(M))
         stack(t.transforms.value, eachslice(reshaped_y; dims=dims))
     end
+end
+function (t::ProductVecInvTransform{<:Elementwise{TypedIdentity},Nothing,Dims{0}})(
+    y::AbstractVector
+)
+    return reshape(y, t.transforms.size)
 end
 
 function with_logabsdet_jacobian(
@@ -139,4 +198,10 @@ function with_logabsdet_jacobian(
         x = stack(lj, eachslice(reshaped_y; dims=dims))
         x, lj.logjac
     end
+end
+function with_logabsdet_jacobian(
+    t::ProductVecInvTransform{<:Elementwise{TypedIdentity},Nothing,Dims{0}},
+    y::AbstractVector{T},
+) where {T}
+    return reshape(y, t.transforms.size), _fzero(T)
 end