+++ HISTORY.md
@@ -1,3 +1,15 @@
+# 0.15.21
+
+Add compatibility with Roots.jl v3.
+
+# 0.15.20
+
+Export `Bijectors.VectorBijectors.has_constant_vec_bijector` to allow downstream packages to overload it.
+
+# 0.15.19
+
+Optimise performance of vector bijectors for products of univariate distributions.
+
 # 0.15.18
 
 Implemented vector bijectors for `Bijectors.OrderedDistribution`.